refactor: Move paginações para arquivo especifico
Também cria paginação padrão de uso geral
